Add support for custom Host IDs
As docker containers do not share many of the same unique identifiers as
their host (for instance MAC address), the way that snapclient decides
on the Host ID for a given client is flawed, in that an update to your
container would have the possibility of changing this identifier, and
thus breaking your snapserver config.

As such, this commit adds support for an additional docker environment
variable "CLIENT_HOST_ID" so that the user can decide on an appropriate
identifier, such that the snapserver config is always able to reference
devices correctly.